refactored Register component
