+++ src/components/Form_Components/Register/defaults.js
@@ -1,0 +1,28 @@
+export const default_user = {
+  email: "",
+  password: "",
+  first_name: "",
+  last_name: ""
+};
+
+export const default_values = {
+  password: {
+    error: false,
+    helperText: ""
+  },
+  email: {
+    error: false,
+    helperText: ""
+  },
+  first_name: {
+    error: false,
+    helperText: ""
+  },
+  last_name: {
+    error: false,
+    helperText: ""
+  },
+  button: {
+    disabled: false
+  }
+};
